Review versioning to make >= 5.3 work again
