+++ src/BatchRunner.php
@@ -5,26 +5,11 @@ namespace WebChefs\QueueButler;
 // Package
 use WebChefs\QueueButler\Contracts\AbstractBatchRunner;
 
+// Framework
+use Illuminate\Queue\WorkerOptions;
 
 class BatchRunner extends AbstractBatchRunner
 {
-
-    /**
-     * Listen to the given queue in a loop.
-     *
-     * @param  string  $connectionName
-     * @param  string  $queue
-     * @param  \Illuminate\Queue\WorkerOptions  $options
-     * @return void
-     */
-    public function daemon($connectionName, $queue, WorkerOptions $options)
-    {
-        $this->validOptions($options);
-
-        // For daemon to use batch if called by accident.
-        // $this->batch($connectionName, $queue, $options);
-        parent::daemon($connectionName, $queue, $options);
-    }
 
     /**
      * Run parent daemon withe the correct args for Laravel version.
@@ -38,7 +23,32 @@ class BatchRunner extends AbstractBatchRunner
      */
     protected function runDaemon($connectionName, $queue, BatchOptions $options)
     {
-        $this->daemon($connectionName, $queue, $options);
+        $this->daemon($connectionName, $queue, $options->workerOptions);
+    }
+
+    /**
+     * Determine if the batch should process on this iteration.
+     *
+     * @return bool
+     */
+    protected function daemonShouldRun(WorkerOptions $options)
+    {
+        $this->checkLimits();
+        return parent::daemonShouldRun($options);
+    }
+
+    /**
+     * Raise the after queue job event.
+     *
+     * @param  string  $connectionName
+     * @param  \Illuminate\Contracts\Queue\Job  $job
+     * @return void
+     */
+    protected function raiseAfterJobEvent($connectionName, $job)
+    {
+        $this->jobCount++;
+        parent::raiseAfterJobEvent($connectionName, $job);
+        $this->checkLimits();
     }
 
 }
